gibberish

+++ headers/Assasin.h
@@ -1,37 +1,38 @@
-#include <iostream>
-#include <string>
-#include <fstream>
 #include <cstdlib>
 #include <ctime>
+#include <fstream>
+#include <iostream>
+#include <string>
+
 #include "Character.h"
 using namespace std;
 
 #ifndef ASSASIN_H
 #define ASSASIN_H
 
-class Assasin : public Character // Derived Class
+class Assasin : public Character  // Derived Class
 {
-public:
-    Assasin();
-    Assasin(string, int, int, int, int, int, int, int, int, int);
-    ~Assasin();
+ public:
+  Assasin();
+  Assasin(string, int, int, int, int, int, int, int, double, int);
+  ~Assasin();
 
-    void setCrit(int);
+  void setCrit(int);
 
-    int getCrit();
+  int getCrit();
 
-    // Actions
-    void attack(Character &);
-    void critAttack(Character &);
-    // void defend(Character &) override;
-    void battleMenu(Character &);
-    void battleMenuBot(Character &);
-    void displayInfo();
+  // Actions
+  void attack(Character &);
+  void critAttack(Character &);
+  // void defend(Character &) override;
+  void battleMenu(Character &);
+  void battleMenuBot(Character &);
+  void displayInfo();
 
-    void save();
+  void save();
 
-private:
-    int crit;
+ private:
+  double crit;
 };
 
 #endif
